feat(handler): Add support for QNAP networking device firmware
QNAP networking devices use the same PC1 cipher as NAS firmware but
derive the decryption key directly from the last 74 bytes of the
encrypted firmware itself rather than a hardcoded secret prefix.
Refactors shared cipher and struct definitions into _qnap.py.
