+++ tests/integration/archive/qnap/qnap_networking/__input__/sample.img
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:6abee30d6fbb0576f8b1d20f8977ef5bd266f9ef5df591c820f523a14db18115
+size 219
