+++ tests/integration/archive/qnap/qnap_networking/__output__/sample.img_extract/sample.img.decrypted
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:b3e506dd3bf06bcb0f4734c3a20cb578ded8d53faa2df6a8f14393f17a23d637
+size 145
