+++ tests/integration/archive/qnap/qnap_networking/__output__/sample.img_extract/sample.img.decrypted_extract/gzip.uncompressed
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:9913f7db663926c49a78de817cd878c21c325303666cbd749c31566d7735f50a
+size 10240
