+++ tests/integration/archive/qnap/qnap_networking/__output__/sample.img_extract/sample.img.decrypted_extract/gzip.uncompressed_extract/testfile.txt
@@ -1,0 +1,3 @@
+version https://git-lfs.github.com/spec/v1
+oid sha256:2035bae46e4534401cf0ad56e38524502304b99e9292e6a35e0eae1667848a0c
+size 21
